+++ src/NextUnit.Generator/Helpers/DataSourceAttributeReader.cs
@@ -371,7 +371,7 @@ internal static class DataSourceAttributeReader
             return DataSourceMemberKind.Unknown;
         }
 
-        foreach (var member in typeSymbol.GetMembers(memberName))
+        foreach (var member in DataSourceMemberResolver.GetCandidateMembers(typeSymbol, memberName))
         {
             if (!member.IsStatic)
             {